Move ExecConfig to types.
