+++ api/server/router/container/backend.go
@@ -14,7 +14,7 @@ import (
 
 // execBackend includes functions to implement to provide exec functionality.
 type execBackend interface {
-	ContainerExecCreate(config *runconfig.ExecConfig) (string, error)
+	ContainerExecCreate(config *types.ExecConfig) (string, error)
 	ContainerExecInspect(id string) (*exec.Config, error)
 	ContainerExecResize(name string, height, width int) error
 	ContainerExecStart(name string, stdin io.ReadCloser, stdout io.Writer, stderr io.Writer) error
