+++ daemon/exec_unix.go
@@ -3,14 +3,14 @@
 package daemon
 
 import (
+	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/container"
 	"github.com/docker/docker/daemon/execdriver"
-	"github.com/docker/docker/runconfig"
 )
 
 // setPlatformSpecificExecProcessConfig sets platform-specific fields in the
 // ProcessConfig structure.
-func setPlatformSpecificExecProcessConfig(config *runconfig.ExecConfig, container *container.Container, pc *execdriver.ProcessConfig) {
+func setPlatformSpecificExecProcessConfig(config *types.ExecConfig, container *container.Container, pc *execdriver.ProcessConfig) {
 	user := config.User
 	if len(user) == 0 {
 		user = container.Config.User
